stuff that wasn't fun, but I sort of started
