+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".Settings" />
+        <activity android:name=".UserSettings" />
+        <activity android:name=".AppSettings" />
         <activity
             android:name=".LoginActivity"
             android:label="@string/title_activity_login">
@@ -29,6 +30,7 @@
             android:name=".MainActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar" />
+
         <service
             android:name="com.backendless.push.BackendlessFCMService"
             android:enabled="true">
