+++ app/src/main/res/xml/settings_app.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <CheckBoxPreference
+        android:key="notification_switch"
+        android:defaultValue="false"
+        android:title="Check box preference" />
+</PreferenceScreen>
